refactor: typo 및 유틸 폴더로 이동
