+++ src/utils/colorSelector.ts
@@ -1,8 +1,8 @@
-import DEFAULT_COLORS from '../../../themes/colors';
+import DEFAULT_COLORS from '../themes/colors';
 
 export type Variant = 'default' | 'primary' | 'success' | 'warning' | 'error';
 
-export const getLighterColor = (variant?: Variant) => {
+export const get400Color = (variant?: Variant) => {
   switch (variant) {
     case 'default':
       return DEFAULT_COLORS.gray_400;
@@ -19,7 +19,7 @@ export const getLighterColor = (variant?: Variant) => {
   }
 };
 
-export const getDarkerColor = (variant?: Variant) => {
+export const get500Color = (variant?: Variant) => {
   switch (variant) {
     case 'default':
       return DEFAULT_COLORS.gray_500;
@@ -33,5 +33,22 @@ export const getDarkerColor = (variant?: Variant) => {
       return DEFAULT_COLORS.error_500;
     default:
       return DEFAULT_COLORS.gray_500;
+  }
+};
+
+export const get600Color = (variant?: Variant) => {
+  switch (variant) {
+    case 'default':
+      return DEFAULT_COLORS.gray_600;
+    case 'primary':
+      return DEFAULT_COLORS.primary_600;
+    case 'success':
+      return DEFAULT_COLORS.success_600;
+    case 'warning':
+      return DEFAULT_COLORS.warning_600;
+    case 'error':
+      return DEFAULT_COLORS.error_600;
+    default:
+      return DEFAULT_COLORS.gray_600;
   }
 };
